pyproject.toml
[tool.black]
line-length = 88
target-version = ["py38"]

[tool.poetry]
name = "idasen"
description = "ikea IDÅSEN desk API and CLI."
authors = ["Alex M. <alexmgit@protonmail.com>"]
version = "0.1.0"
license = "MIT"
readme = "README.rst"
repository = "https://github.com/newAM/idasen"
homepage = "https://github.com/newAM/idasen"
keywords = ["ikea", "idasen", "bluetooth", "linak", "ble"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
]

[tool.poetry.dependencies]
python = "^3.8"

bleak = "~0.7.1"
pyyaml = "^5.3.1"
voluptuous = "~0.11.7"

[tool.poetry.dev-dependencies]
black = { version = "20.8b1", allow-prereleases = true }
coveralls = "^2.0.0"
flake8 = "^3.8.3"
flake8-bugbear = "^20.1.4"
pep8-naming = "^0.11.1"
pytest = "^6.0.1"
pytest-asyncio = "~0.14.0"
pytest-cov = "^2.10.1"
sphinx = "^3.1.1"
sphinx-rtd-theme = "~0.5.0"
toml = "^0.10.1"

[tool.poetry.scripts]
idasen = "idasen.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
